# swagger.yaml
openapi: 3.1.0
info:
  title: Trading Platform Backend API
  description: |
    A comprehensive trading platform backend with IIFL integration, real-time market data, 
    portfolio management, and advanced stock analysis features.
    
    ## Features
    - **Google OAuth Authentication** - Secure user authentication
    - **IIFL Trading Integration** - Place orders, manage positions
    - **Real-time Market Data** - Live prices, quotes, and market analytics
    - **Portfolio Management** - Track P&L, holdings, risk metrics
    - **Stock Analysis** - Advanced analytics and returns calculation
    - **WebSocket Support** - Real-time updates
    
  version: 1.0.0
  contact:
    name: Trading Platform Support
    email: support@tradingplatform.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://trading-backend-oab.info
    description: Production server
...
paths:
...
  /api/returns/files:
    get:
      summary: List stock returns files
      description: Get list of available stock returns data files from S3
      tags:
        - Stock Returns
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Files list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnsFilesListResponse'
  /api/returns/file/{filename}:
    get:
      summary: Get returns file data
      description: Download and parse a specific returns data file
      tags:
        - Stock Returns
      security:
        - BearerAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          example: "adjusted-eq-data-2024-01-15.csv"
      responses:
        '200':
          description: Returns file data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnsFileDataResponse'
  /api/returns/all:
    get:
      summary: Get all stock returns
      description: Get returns data for all stocks from the latest file
      tags:
        - Stock Returns
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All stock returns retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockReturnsListResponse'
  /api/returns/{symbol}:
    get:
      summary: Get stock returns by symbol
      description: Get historical returns and scores for a specific stock
      tags:
        - Stock Returns
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          example: "RELIANCE"
      responses:
        '200':
          description: Stock returns retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockReturnsResponse'

  # ========== STOCK ANALYSIS ENDPOINTS ==========
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from Google OAuth login.
        
        **How to authenticate:**
        1. Login via `/api/auth/oauth/google/login`
        2. Complete OAuth flow
        3. Receive `access_token` and `refresh_token`
        4. Use `access_token` in Authorization header: `Bearer {token}`
        5. Refresh token when expired using `/api/auth/refresh`
  schemas:
...
    StockReturnsData:
      type: object
      properties:
        symbol:
          type: string
        fincode:
          type: string
        isin:
          type: string
        latest_date:
          type: string
          format: date-time
        latest_close:
          type: number
        latest_volume:
          type: integer
        turnover:
          type: number
          nullable: true
        returns_1_week:
          type: number
          nullable: true
        returns_1_month:
          type: number
          nullable: true
        returns_3_months:
          type: number
          nullable: true
        returns_6_months:
          type: number
          nullable: true
        returns_9_months:
          type: number
          nullable: true
        returns_1_year:
          type: number
          nullable: true
        returns_3_years:
          type: number
          nullable: true
        returns_5_years:
          type: number
          nullable: true
        raw_score:
          type: number
          nullable: true
          description: Composite performance score
    StockReturnsResponse:
      type: object
      properties:
        status:
          type: string
        symbol:
          type: string
        data:
          $ref: '#/components/schemas/StockReturnsData'
        source_file:
          type: string
        timestamp:
          type: string
    StockReturnsListResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/StockReturnsData'
        total_count:
          type: integer
        source_file:
          type: string
        timestamp:
          type: string
    ReturnsFilesListResponse:
      type: object
      properties:
        message:
          type: string
        files:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
              s3_key:
                type: string
              size_mb:
                type: number
              last_modified:
                type: string
              source:
                type: string
        total_files:
          type: integer
        source:
          type: string
        timestamp:
          type: string
    ReturnsFileDataResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/StockReturnsData'
        total_count:
          type: integer
        source_file:
          type: string
        file_size_mb:
          type: number
        last_modified:
          type: string
        source:
          type: string
        timestamp:
          type: string

    # ========== BHAVCOPY SCHEMAS ==========
...
security:
  - BearerAuth: []